Add version of FEInterface::n_dofs() taking an Elem *.
This version should be preferred to the other version of
FEInterface::n_dofs() where the user is responsible for figuring out
the true order of the element (including p-refinement) in cases where
an Elem pointer is available. Unfortunately we can't easily deprecate
the other version, since an Elem pointer is not always available.