distcheck: add new tecio.a symbolic link
I think because of the update to automake 1.16.1, build-aux/config.guess
is now reporting that my system is "x86_64-pc-linux-gnu", so I needed to
add a new symbolic for the Tecplot binaries. Note that you have to
force add (add -f) these links because otherwise they are ignored by
the "lib" line in .gitignore.