Handle different combinations of METIS/ParMETIS in PETSc.
The basic idea is that we don't want to mix building our own
METIS and using PETSc's ParMETIS. The different possibilities are:

                       | Parmetis in PETSc?
                       |   yes  |    no
                   ----------------------
                   yes |    A   |    B
METIS in PETSc?    ----------------------
                   no  |    C   |    D

A: PETSc was built with both METIS and ParMETIS support. In this case,
   we use both from PETSc.
B: PETSc is built with METIS support but no ParMETIS support. In this
   case, we use the METIS support and disable ParMETIS.
C: PETSc is built with no METIS support, but does have ParMETIS. I'm not
   sure this is really possible, but in this case, we would build our
   own METIS and disable ParMETIS support just to avoid mixing them.
D: PETSc is built with neither METIS nor ParMETIS support. In this case,
   we build both on our own.