ExodusII_IO: Fix write_element_data_from_discontinuous_nodal_data()
There was an issue with subdomain restricted variables. I previously
did not realize that they would still show up in the discontinuous
solution vector (as zero padding) so I was not indexing correctly in
that case. This fix should also be cherry-picked onto v1.4.0 since
this API is present in that version.

Refs #2029.
Refs #2010.