Set TAR_OPTIONS to improve portability of tarballs.
Instead of setting TAR_OPTIONS, I also tried calling
AC_SUBST(am__tar,[tar --hard-dereference]) but unforunately this
alternate approach did not work, perhaps the AC_SUBST comes too late
in the process to take effect or perhaps it is simply not possible to
do this. It was an off-hand remark in a mailing list [0], not an
officially sanctioned way of doing things.

[0]: http://gnu-automake.7480.n7.nabble.com/bug-19616-dist-tarball-contains-hardlinks-td21667.html

Refs #2046.
Refs #2029.