Change error message when constraint loops are detected.
As discussed in the documentation accompanying this PR, the "cyclic
constraint" terminology is used by some software to indicate a certain
type of rotational constraint, and we don't want our error message to
be confused with this. Both APIs can remain in place for a while but
eventually DofMap::check_for_cyclic_constraints() will be deprecated.