Add DofMap::reinit_send_list()
This is an API hook to reinitialize the send list. This may be
necessary for example when a user can only add an algebraic or
coupling GhostingFunctor after the DofMap has been initialized.
Such scenarios arise then the GhostingFunctor requires the evaluation
of solution fields in order to determine ghosting/coupling.