Drop VecScatterCreateWithData() workaround.
This function did not wind up making it into any PETSc release.  This
should probably also be cherry-picked onto the v1.4.1 branch. Even
though the likelihood of any tagged libmesh release continuing to work
with PETSc master is negligibly small, the tagged release _might_ work
with a future PETSc 3.11 so we should try and support that if
possible.

Refs #2084.