Fix remote_elem copying in UnstructuredMesh
We've never been correctly copying remote_elem neighbor links in the
skip_find_neighbors code path (or rather, we've been copying them
correctly and then accidentally *undoing* them), and since #2028
made skip_find_neighbors the default behavior in more code paths, that
bug was starting to be triggered e.g. in the #1938 tests.

This commit gets remote_elem into the old-to-new-elements map so the
skip_find_neighbors code path will recognize it, and removes a bit of
redundancy in the rest of the old-to-new-elements map initialization,
and moves the manual remote_elem-specific copying code into the
!skip_find_neighbors path to avoid redundancy with that.