Don't constrain hanging nodes at variable boundary
If we have a subdomain-restricted variable which is supported on a
fine element but *not* supported on a neighboring coarse element, then
we do *not* want to try to generate hanging node constraints for that
interface.  Doing so was giving us OOB reads and garbage DoF IDs.

Backported with the other commits from #2099 with the command.
git cherry-pick e9297f171 33466d071 ae5cbc117 de07ffc38

Refs #2029.