ExodusII: write all nodesets simultaneously.
This is similar to what we did in b4ac1df1 (#968) to speed up the
writing of multiple subdomains.

For a mesh with about 26k nodesets, this speeds up a test code that
reads in and writes out the mesh from about 700 seconds to about 30
seconds. I think a similar optimization on the reading side would be
able to get that number down even lower.