Use Real in Eigen dense linear algebra
The many typedefs here are good fix for making sure the examples'
precision matches the libMesh configuration precision.

The handful of foo = Real(bar) changes here are ugly workarounds for
the fact that Eigen sometimes returns template expressions that it
expects users to implicitly convert to scalars, boost float128
operator= accepts scalars that it knows how to implicitly convert to
float128, and C++ compilers balk at doing multiple implicit conversion
steps in a row.  Not sure what a better fix would be.