Add unit test for Legendre radial basis functions.
The radial basis functions are not used directly, but they are used in
the computation of the phi and dphi values for infinite elements. In
this test, we tabulate several arbitrarily selcted phi, dphi values in
order to verifty that changing the legendre_eval() implementation does
not change these values.