Generalize subdomain comparison for var groups
This allows user code to be more general. They can create
a std::set instance in all cases and pass it in; if the set
is empty then this code will do the right thing and
recognize that there are no restrictions. The
their_active_subdomains was already doing the right thing
because it is a nullptr if the previously added subdomain
set was empty (e.g. see Variable::implicity_active())