Fixes to get libmesh compiling with PETSc master as of late July 2019.
* Fix for SNES_CONVERGED_TR_DELTA.
* Fix for dm->ops->getinjection.
* Fix for dm->ops->getaggregates.

All of these changes are currently in PETSc master but not yet in any
PETSc release.  For example, the SNES_CONVERGED_TR_DELTA was merged
into PETSc master 16 Jul 2019 in: c1c6be5a2e, and the latest PETSc
release is v3.11.3 (26 Jun 2019).  These changes are therefore likely
to be in 3.12 but not in any 3.11.x release. Also, this change will
break anyone who is using libmesh with an older version of PETSc
master, but there is basically nothing we can do about that, they just
have to update.