Make mallocs uniformly an error in PetscMatrix
This adds the setting to the other two init methods that
if a new malloc occurs during MatSetValues then it is an
error. I think this is an improvement because it establishes
uniformity across our init methods and it can prevent users
from running extremely slow simulations.

I fully expect this to cause failures in MOOSE...