More digits for SEVENTH-order quadrature on TRIs.
This is related to the work in #2185, which enables higher than double
precision calculations. It does not look like the quadrature unit test
max order for triangles was changed in that PR, so I guess the previous
rule was sufficient for O(TOLERANCE^1.5) quadrature accuracy. Perhaps
with the new more accurate parameters we could use an even tighter
tolerance in testTriQuadrature()?