Add ExodusII_IO::write_sideset_data()
Variables (data) on sidesets are conceptually similar to elemental
data fields defined on lower-dimensional subdomains of elements.  They
can be useful for defining boundary condition data, such as
e.g. piecewise-constant normal forces on each face of a sideset.

Since we don't normally store such data on Systems, I just invented a
data structure which was fairly convenient for my purposes.  This API
isn't written in stone, though, we can change it to something else
before a bunch of backwards-incompatible code has a change to be
written, if desired.