Refactor ExodusII_IO_Helper classes
Primarily the ElementMaps and Conversion classes are updated. This is
in preparation for adding support for reading/writing edgesets and
associated data fields to Exodus files. When this code was originally
written, we used C arrays because std::vectors could not be
initialized using initializer lists (indeed, no such thing existed at
the time). This required us to pass array sizes to several different
interfaces, which made them bloated and ugly. This refactoring gets
ride of some of that ugliness.