Call newer exII::ex_put_concat_all_blocks() API.
This API will eventually allow us to specify parameters for blocks of
edges and faces to be written to the Exodus file. Note that we have to
add dummy 'num_{edges/faces}_per_elem' fields even though we are not
currently using them, otherwise the code segfaults, probably while
trying to access into these arrays...

It seems like we should be able to set define_maps==1 when writing
block parameters.  This is an optimization that pre-allocates space
for the node_num_map and elem_num_map fields if you are going to write
them.  It seems that we always write these, so we should take
advantage of this optimization, but when I tried to turn on this flag,
I got an error:

Exodus Library Error: [ex_put_concat_all_blocks]
	Error: failed to find node map size of file id 65536
	[-46] Invalid dimension id or name
Error writing element blocks.

I think the issue is that we have to first specify the number of
elem_num and node_num maps that will be written during the call to
ex_put_init_ext(), so I will try that after I make sure all the MOOSE
testing is also happy with this change.