Keep track of FE requests for mapping data
Before, if *only* mapping data was requested, we wouldn't realize that,
we would think that nothing had been requested, and we would calculate
all data at reinit(), for backwards compatibility.  This was making the
Elem::volume() fallback code too slow in the general case, and was
breaking it (due to unimplemented second derivatives) in the Rational
Bezier case.