Conversion class: handle identity maps more elegantly.
We now treat a missing node/edge/face map as the "identity" or
"identity+1" map as appropriate.  This allows us to save a lot of
lines of boilerplate arrays, etc. and simplifies the construction of
the Conversion objects.

Note that every Exodus element except the Hex27 has an identity node
map, so by not storing identity node maps we can reduce the amount of
data required by quite a bit.

Drop hex27_face_map and inverse, they were the same as hex_face_map.