In Elem::level check interior_parent->level()
In MOOSE mortar when building the mortar segment mesh,
we set the interior_parent for the mortar segments, but
we do not set the parent, because parent implies the child
was formed through adaptivity. However, when a mesh has been
refined and the interior_parent has been refined, then there
are failed assertions in debug mode during prepare_for_use
because the interior_parent level != 0 while previously if the
parent was nullptr the mortar segment would report as being level 0.
Hence I think it makes sense to check the interior_parent level
if the parent is nullptr.