When building node list from side list, first reset node containers
Here resetting means to reset the node list to be that obtained
only from calling add_node with from_side = false, e.g. it should
represent nodes that were generated from nodesets.
I had to do this for some GeoemetricSearch work
in MOOSE combined with mesh coarsening. The node list was containing
some nodes that did not exist on the coarsened mesh and it was
because the BoundaryInfo node containers were not getting cleared