Additional options for adding extra integers+data
We now go back and allocate space on nodes and elements in a non-empty
mesh by default.  So that we don't do this O(N) times when adding N
integers, this feature can be manually disabled, or it can be used in
O(1) fashion by using the new vectorized APIs.