namespace libMesh::Parallel -> PassMess
This isn't as smooth as I'd hoped it would be.  The shim class
libMesh::Parallel::Communicator is an ugly hack, but it's the only way
I've figured out to stay compatible with forward declarations.  And
there doesn't seem to be any way to convince a compiler that, when it
sees a partial template specialization of libMesh::Parallel::Packing,
that's *really* a specialization of PassMess::Packing, so I'm leaving
Packing in the original namespace for the moment.