Use ReplicatedMesh in new edgeset unit test
Otherwise we get a failure from --enable-parmesh builds, because
something about either this test or the underlying code isn't liking
DistributedMesh.

This is a short-term workaround so that I don't again run into what
turned out to be an astonishingly hard bug to track down.

As a medium-term fix I'm hoping @jwpeterson can figure out that "isn't
liking DistributedMesh" problem and then we can revert this.

As a long-term improvement, see #1307.  We need something better than
CppUnit for parallel tests, to avoid future "astonishingly hard bug to
track down" issues.  This was another #1008 case, and because I was
working in a branch I had to bisect with --re just to determine that it
really was a new test failing rather than a bug in my branch.