Use DirichletBoundary in example
I left some examples using penalty conditions, way back when, to
demonstrate that alternative, but we've got more penalty formulations
than we need in the vector examples these days.  Let's make this one
more "normal", both to get better exercise of boundary projections and
to make it more useful for benchmarking.

(In reality I made this change to explore a solver convergence issue
at high p, and that turned out to be a red herring, but I still think
this is an improvement)