Add BoundaryInfo::erase_if() helper function.
You have to be a bit careful when erasing from maps in general, to
avoid invaldating the iterator you are using. With multimaps you also
sometimes have to iterate over ranges of potential erasure
candidates. Instead of copy/pasting the code which does this, we can
have a generic helper function that takes a lambda.  IMO this is still
a bit difficult to read due to the decltype bits, but it will be
improved when we can use generic lambdas in C++14.