Sync nodes after all_second_order if not replicated
Previously we only synced if we were not serial. However,
even if we haven't deleted remote elements in our distributed mesh
we still need to sync the nodes because even if we are adding
nodes on all procs they are getting temporary out-of-sync ids.
If we don't do this sync, then I get failed assertions
in libmesh_assert_valid_parallel_ids. Explicitly, even for the very
first new node, I can fail (min_proc_id == this->processor_id() && obj)
because the non-owning proc may have assigned the node
ID 14 while the owning proc assigned it ID 17 or whatever. Then
the owning proc will fail the assertion when `i == 14` because
the LHS will be true while obj is nullptr