Use lower-order quadrature in JumpErrorEstimator
We can get nearly-as-good results at roughly half (2D) or a quarter (3D)
of the cost by not overintegrating.

We'll have to make this non-default behavior to avoid breaking people's
regression tests and reproducability though.  IMHO the savings are worth
it regardless, but any change in behavior this significant probably
deserves to be opt-in.