Use relative URL for TIMPI submodule.
Technically we could use a single ../ but I think this helps make it
more clear that the TIMPI repo is located under the libMesh project.

If you already pulled a version with the previous URL, you may need to
run the command:

git submodule sync contrib/timpi

in order for the update URL to take effect.

Refs #2386.