Update to TransientRBConstruction to not use lapack calls directly
Previously we made direct calls to lapack in TransientRBConstruction::enrich_RB_space()
which led to code that is hard to understand. This commit replaces the lapack calls
with calls that use the DenseMatrix API. The functionality is identical, but the code
is more readable.

Small fix to print statement

Whitespace updates in RBConstruction to follow libMesh conventions