eval_mixed_derivatives, eval_at_point skip_context
If we're evaluating the off-center points in a finite differencing
stencil for HERMITE cross derivatives, and if we had to fully evaluate
the center point first, we don't need to do a context check (because the
old context is good even if we're not quite on the same old element),
and we don't want to do a context check (because we can get a false
positive assertion failure when the code sees that we're not on the same
old element even though we're on the same new element and the new
element wasn't JUST_COARSENED).

If we were able to just copy DoF coefficients at the center node, then
we can skip the finite differencing entirely, and just copy the cross
derivatives too.  The API for this is a little hackish, but its use
replaces the worse finite differencing hack, resulting an a net
*decrease* in hackishness.