Add ParallelType to ImplicitSystem::add_matrix API
This mirrors what is done for System::add_vector. We also keep
track of the matrix types in a _matrix_types data member. This
also adds ParallelType as a default argument to the formerly
zero arg SparseMatrix::init method. For almost all concrete
SparseMatrix data types this parameter will be unused, but its
useful for DiagonalMatrix and it could be useful for other derived
matrix types in the future as it is useful for vectors now.