Add MeshRefinement::add_elem(unique_ptr)
As the API implies, this calls the new MeshBase::add_elem(unique_ptr)
internally, and is intended to make it clear that the MeshRefinement
object (actually the Mesh inside it) takes ownership of the Elem
pointer that gets passed in.

Call new version of MeshRefinement::add_elem() in Elem::refine().