Use ghosted projected_solution in URefEE
We were already only localizing the ghost DoFs, but that means there's
no point in allocating space for all the others too.  Using ghosted
vectors shouldn't cost too much CPU time (or might be a win depending on
cache and heap behavior) and should be a huge win in parallel in terms
of memory usage.