BoundaryInfo::add_elements(): Add flag controlling setting of parent side ids
We always store parent side ids when calling BoundaryInfo::sync(), but
by default, we don't store this information on direct calls to
BoundaryInfo::add_elements(). The reasoning is that we don't want to
pay for this extra storage on _all_ elements when it only really makes
sense for lower-dimensional elements. The user can override this
default by passing "true" for the store_parent_side_ids flag.

Refs #2457