Forward declare BoundaryInfo in mesh_base.h
As far as I know, you can forward declare the template argument of a
std::unique_ptr provided that the containing class' destructor is
declared out-of-line, otherwise knowledge of how to delete the pointed-to
object is required.

This may cause some temporary compilation failures in downstream apps,
but I think it's worth it if we can avoid recompiling basically the
entire library whenever boundary_info.h is touched.