Add DirichletBoundary copy assignment operator.
Also add defaulted destructor, move constructor, and assignment
operators.  As far as I understand it, you can std::swap a class T
which is both move-constructible and move-assignable, but I'm not sure
if this is guaranteed to work in C++11 or if some later standard is
required. If it works for our "min" compilers, then I'm happy.

https://en.cppreference.com/w/cpp/algorithm/swap