Add 'extra_checks' parameter to FEMap::inverse_map()
Previously, the 'secure' boolean was used to control both the
convergence checks within each Newton iteration and whether additional
"integrity" checks were performed once the Newton iteration was
considered to have converged.

In certain cases, it is useful to still have the Newton
convergence/divergence checks enabled, but to skip the additional
integrity checks that take place once the Newton scheme is considered
to have converged. By default, the extra_checks flag is true (same as
the existing 'secure' flag).