Small fix to RBConstruction
We recently made a small change to train_reduced_basis to break out
if we encountered a zero solution. This is normally fine, but it can
cause an issue in the case that we have use_empty_rb_solve_in_greedy=false,
such as in RBEIMConstruction. This is a fix for that case.