Add templated make_range() helper functions.
The existing index_range() helper function could be implemented in
terms of make_range() instead, i.e.

for (auto i : index_range(v))

would change to:

for (auto i : make_range(v.size()))

so it's a bit more typing, but the advantage is that we don't have to
make an index_range() specialization for all possible types of v.