Switch to C++11 scoped enumerations.
This is the right solution if you want to have multiple enumerations
defined in the same class with the same name. I'm not sure if it's
backwards compatible to switch from an "int" enum to a scoped enum,
but hopefully any user code was already prefixing the use of these
enums and will be unaffected.

Another option would be to change the name of the existing
NodeBCTupleSortBy enumeration (BCTupleSortBy?) and then append ELEM_ID
and SIDE_ID enumerations to the list, but that is _definitely_ not
backwards compatible.

A final option is to append the new enumerations to the list but not
change the name, but that just seems confusing at best.