Don't undef or define MPI_REAL
MPI_REAL is part of the MPI standard, and trying to redefine it is
bad.

I've always been a bit dubious about this, but I've finally found an
MPI implementation with which it breaks, with a "redefined macro"
warning that dies on --enable-werror.

Fortunately, we seem to have long since stopped using MPI_REAL in
libMesh, in favor of TIMPI::StandardType autodetection and such.  I
don't see any uses in MOOSE or GRINS either.