Initial refactoring of ConstrainDirichlet class
The eventual goal is to change the order of the loops from
elem-loop-innermost:

for (dirichlet : dirichlets)
  for (var : dirichlet->variables)
    apply_dirichlet_impl<T>(range, var, *dirichlet, ...);
      for (elem : range)

to an order in which there is only one total loop over the elems:

for (elem : range)
  for (dirichlet : dirichlets(elem))
    for (var : dirichlet->variables)
      apply_dirichlet_impl<T>(elem, var, variable, fe_type, *dirichlet);

This new approach should work better in cases where there are many
DirichletBoundary objects, but each one is defined for only a few (or
1) nodes or elements, since we will go from N loops over the elements
down to 1.

This initial round of refactoring does not yet re-order the loops, but
makes the code a bit more modular by employing a FEMContext
object. The FEMContext is used to store all the required elem, side,
and edge FE objects, quadrature rules, DOF indices, etc.

Note: the FEMContext used in ConstrainDirichlet is assigned an
algebraic_type of DOFS_ONLY because we aren't allowed to access
current_local_solution vector values when ConstrainDirichlet is
called...since they may not be initialized yet.