Use std::sort() in Elem::compute_key()
Since we are building a fixed-size array to pass to the hashword()
function anyway, might as well call std::sort instead of using a
hand-coded version.

We also no longer assert that the node ids are distinct in addition to
being sorted. This makes the function a bit more flexible in some
non-standard cases (degenerate Elems).