Add apply_lagrange_dirichlet_impl() optimization
This is the third part of the DirichletBoundary refactor
(see #2567, #2574, #2583 for the other parts). This part adds an
optimization for the important special case of specifying BCs for
Lagrange variables.

The basic idea is that Lagrange FEs only have nodal DOFs, so we should
only have to interpolate the boundary function -- projection should
not be required. To implement this, we needed to add some additional
code for shellface BCs which now induce nodal BCs on all the nodes of
the shell element.

If there's no DOFs, there's nothing to constrain.