Respect original order of DirichletBoundary objects.
The refactoring in #2574 caused the DirichletBoundary objects'
constraints to be applied in an "arbitrary" order rather than the
order in which they were actually added to the DofMap.  Since users
may have been implicitly or explicitly relying on this ordering to
break "ties" in cases where multiple constraints were applied to the
same DOF, we should try to maintain that order, even though it was not
guaranteed previously. I also updated the documentation to make this
ordering guarantee an official part of the DofMap API.

* Add 'boundary_id_to_ordered_dirichlet_boundaries' which is ordered
  based on the original vector ordering.
* Add 'ordered_dbs' container to SingleElemBoundaryInfo struct.
* Update comment for DofMap::add_dirichlet_boundary()