Actually fix the constraints bug!
It turns out that, without ghosting of all point neighbors, we can't
trust find_point_neighbors to return a complete set for partition
boundary points on a distributed mesh, which means we can't
necessarily look at a complete set of element ids to disambiguate
which side should constrain which.

Looking at the specific points themselves to disambiguate constraints
fixes the problem for me, at least on the case where it manifested,
systems_of_equations_ex9.C with 21 processors.