Use FEInterface::n_dofs_at_node() in FEGenericBase::coarsened_dof_values()
The new code mimics the original code in which the "temp_fe_type"
variable did not take into account the current Elem's p_level(). I'm
not sure whether this is correct, but at least it's consistent.

If necessary, we can initialize extra_order to elem->p_level() instead
to fix this issue.