compute_node_constraints(): ignore elem->p_level() inherited by sides.
I'm not sure why this passed testing in #2593 when it was merged, but
the basic idea here is that a Side is built for a potentially
p-refined (say, Hierarchic) element and it (now) inherits the parent
Elem's p_level(). We however want to subsequently create a
"mapping_family" FE on this side, which typically means
LAGRANGE. Therefore if the side came from a (FOURTH, HIERARHIC) Elem
and we continue to use this same p_level(), we are going to run into
issues. This change therefore restores the previous behavior of this
function which was updated in f707c65a, but which turns out to have
been correct all along.