Elem::simple_build_side_ptr returning unique_ptr
For this case, too, we can handle everything with identical side types
using generic code.

In this commit (as well as the ones to follow), we initially set the
"interior" element as a parent() of the side, so that the Elem
constructor will copy things like p level and subdomain id from the
interior, but then we make sure that the interior_parent() and *not* the
parent() link point to the interior element.  This is analogous to the
changes we made for use with boundary meshes, and here it prevents the
code in set_p_level() from messing up interior element p levels.

This fixes a regression in adaptive hp refinement caused by #2592