build_side_ptr(proxy) parent() w/deprecated
Now if we haven't configured with --disable-deprecated, the current
build_side_ptr behavior will match the previous build_side_ptr behavior,
which was to leave side->parent() pointing to the interior_parent for
proxy sides but leave it null for non-proxy sides.

This is confusing and was surely unintentional, but either setting
parent on non-proxies *or* leaving it unset on proxies causes various
MOOSE unit tests to fail, so let's stay as backwards compatible as
possible until users have enough time to update codes to use
interior_parent() where appropriate.