Check whether nonlinear residual vector and base vector are equal
Reusing the nonlinear residual vector as the base vector for matrix-free
finite differencing was an optimization we introduced in #1572. However,
per idaholab/moose#14644 we would like to be sure that it is a safe
thing to do. So I propose that in debug mode we check and make sure that
they really are the same.

Closes idaholab/moose#14644