Try linking to VTK CommonExecutionModel too
This fixes "undefined reference" failures when linking MOOSE
appliations on some systems.

Not sure what code is looking for vtkAlgorithm and vtkImageAlgorithm
methods, but they appear to be defined in libvtkCommonExecutionModel -
and that library exists in VTK at least back to 6.2, so it should be
safe to pull in regardless.