Add libmesh_error_msg_if(cong, msg) macro
The idea here is to make error checking more self-contained while also
allowing for the possibility of later adding a configure option which
disables _all_ error checking.

Another option would have been to add:
libmesh_always_assert_msg(asserted, msg)

but in that case in order to use the new macro we'd have to reverse
the logic of all the checks that are already present in the code, and
that seemed too error prone to me.