GhostPointNeighbors optimization
This loses us ghosting of different-dimensional point neighbors in
multi-manifold meshes (except those brought in by interior_parent()
links), but that was already unsupported in other code paths.

This should be vastly faster on large processor counts on
ReplicatedMesh, as well as if anyone ever decides to use it for a
coupling matrix.