Better handling of map::lower_bound corner case in SH::find_stored_entry
1) Consider the case where the time for which we were requesting the solution be retrieved is time_of_last_key + FPE.
Then, map::lower_bound wont find a greatest upper bound inside the map and will point to end.
2) We were assuming that map::lower_bound pointing to end meant that we are adding a new entry, but this is only correct
if the map is being populated.
3) Added a 'storing' boolean to the arguments of find_stored_entry, to distinguish between the case where map::lower_bound
pointing to end implies the addition of a new entry, or precipitates an attempt to look for a matching key for time in map entry before end.