std::vector<Real *> QoI_time_instant TS::integrate_adjoint_error_estimate now takes a vector of time instants as a possible argument
1) QoIs in unsteady simulations can either be smooth (temporally distributed) or instantaneous (for example, a final time QoI).
A temporally smooth QoI is assumed by default, but the user can pass in a vector of pointers to time instants corresponding to each QoI.
If any entry in this vector is not NULL, that QoI is assumed to be an instantaneous QoI, and the Real at the pointer location is assumed
to be the instant of evaluation. For such a QoI, the trapezoidal rule cannot be used to time integrate the error estimate, and the left Riemann
sum is used.
2) The variable final time is now initialized to std::nan("1") inside the TS constructor. This allows us to distinguish between a state where the
final time (steady state for example) has not been reached, versus where the time loop has not converged yet.