send_list: handle invalid_id, assert no garbage ids
We need to handle invalid_id in cases where we're running subdomain
expansion problems with subdomain-restricted variables: near the borders
between subdomains we can see undefined DoF indices from the
undefined-variable side.

We need to test for garbage because that's what we were previously
seeing and if we ever encounter that again it's a big problem.