Make our disambiguating constructor explicit
Not sure how we forgot this, but without it a function with an argument
intended to take const Node & can compile when passed dof_id_type
instead, because libMesh will sadistically construct a temporary node
whose x coordinate is the specified id.

Thanks to @rwcarlsen for his finding this and apologies to @rwcarlsen
for our enabling this.