Move ExodusII_IO_Helper::read_nodeset() to Nemesis_IO_Helper
This has now been replaced by ExodusII_IO_Helper::read_all_nodesets().
This was the only function using the node_list variable, so move that
to Nemesis_IO_Helper as well. Technically this was a public member
function, but it was in the Helper class so outside classes should not
have been using it directly...