Use -> to get a member from a pointer
This method is templated and not being instantiated, so I guess my
compiler and our CI compilers didn't bother to check the code's validity
when seeing vector<T> with an indeterminate T, rather than realizing
that a pointer to *any* object can't have a size() method.

This will need to be backported to the release branch, for the sake of
users who want to use this method or who have smarter compilers than
mine.  Thanks to @dschwen for reporting this.