ParsedFEMFunction: change "parsers" to a vector of unique_ptrs
This is an attempt to fix the issue raised in #2710. Since
FunctionParserBase has a deleted move constructor, it is not
considered to be MoveInsertable, which means it cannot be passed to
std::vector::push_back() by value.

Adding a vector<unique_ptr> to the class means that it can no longer
be default copy constructed. We had a unit test of this but, no actual
uses within the library so hopefully it's OK to drop it.