Delete, rather than default, MeshBase move constructor
The MeshBase move constructor can't be safely defaulted because of the
way we construct the `BoundaryInfo` member. Using the default move
constructor, our `BoundaryInfo` in the move-constructed object will
have an invalid reference to its `_mesh` member after the move
constructor method is finished

Also remove the Mesh move constructor unit test that we had previously.

This is a combination of the following two commits cherry-picked from
master:

4a48915ee
681e9adad

Refs #2732