Delete all mesh move constructors
Clang warns of explicitly defaulted, implicitly deleted move
construction in `UnstructuredMesh`, which is definitely true. To remove
those warnings, we explicitly delete all move constructors from
`MeshBase` down

This should have gone in with commit c7cbc0d6b on this release branch,
but it is currently only a warning to default a special function which
is deleted in a parent class, and not all compilers warn about it.