Default/delete special functions for System class
Currently the System class manages memory in the _vectors member, so
none of the special functions can be defaulted. Also, previously we
were doing a bunch of pointer-nullifying in the destructor, which should
only be necessary if other classes hold pointers to those objects and
rely on them being nullified by the System as it goes out of scope for
some reason.