TypeVector overloads for std::norm and friends for Eigen::Matrix
This is needed in MOOSE.  I know this really seems like a moose-specific
hack finding its way into libmesh, but the way moose does unity builds,
we can't guarantee the order of appearance of the declarations of these
overloads.  The compiler keeps finding the wrong ones and ignores the
correct one if we put the overloads anywhere in MOOSE.  I'm happy to
take suggestions if anybody has any idea how to fix this only on the
moose side.