TransientSystem should not store unique_ptrs
Previously, we were creating unique_ptrs out of references handed back
by System::add_vector(), but this seems like it should cause undefined
behavior, as there are now two unique_ptrs to the same object.

There does not seem to be a way to fix this issue without breaking
backwards compatibility, but since we recently tagged a new release
and this apparently fixes a potential bug, this tradeoff seems
worthwhile.