Simplify unique_ptr return statements
If there's no issue with std::unique_ptr<Derived> vs. std::unique_ptr<Base>,
we should be able to simply return the object we've created without any additional
release() gymnastics.

It's possible that this was just a workaround required on older compilers,
so if this works with our current "min" gcc and clang, I think we
should be safe to merge.