Add VecGhostLocalFormRAII helper class
This approach gives better exception safety and avoids some code
duplication because we can do the GHOSTED check within the helper
class itself rather than before every operation on the Vec.

Add LIBMESH_CHKERR2 macro. This is useful when you want to call the
LIBMESH_CHKERR macro but you aren't in a ParallelObject.

In cases where multiple Vecs are involved (e.g. VecAXPY) then even if
only the "primary" Vec (aka the Vec being written to) is GHOSTED,
_all_ Vecs still need to call VecGhost{Get,Restore}LocalForm.
Therefore we have to pass in some additional information to the
VecGhostLocalFormRAII constructor which controls whether or not the
VecGhostLocalForm is created, we can't rely on an individual
PetscVector's type() being "GHOSTED" to tell us what to do.